fix cameraX UI, and set API for expression detection

+++ app/src/main/java/com/ch2ps215/mentorheal/domain/model/Classification.kt
@@ -1,8 +1,6 @@
 package com.ch2ps215.mentorheal.domain.model
 
-import java.util.Date
-
 data class Classification(
     val label: String,
-    val scores: Float,
+    val scores: Int,
 )
